Refactor Name.java class to a separate package
Person, Assignment and Module all have an field for name. Let's move the existing Name class to its own package to be used by all Person, Assignment and Module.
