+++ src/main/java/seedu/edrecord/logic/parser/AddCommandParser.java
@@ -14,9 +14,9 @@ import java.util.stream.Stream;
 import seedu.edrecord.logic.commands.AddCommand;
 import seedu.edrecord.logic.parser.exceptions.ParseException;
 import seedu.edrecord.model.module.Module;
+import seedu.edrecord.model.name.Name;
 import seedu.edrecord.model.person.Address;
 import seedu.edrecord.model.person.Email;
-import seedu.edrecord.model.person.Name;
 import seedu.edrecord.model.person.Person;
 import seedu.edrecord.model.person.Phone;
 import seedu.edrecord.model.tag.Tag;
