+++ src/main/java/seedu/edrecord/logic/parser/ParserUtil.java
@@ -12,9 +12,9 @@ import seedu.edrecord.commons.util.StringUtil;
 import seedu.edrecord.logic.commands.MakeModuleCommand;
 import seedu.edrecord.logic.parser.exceptions.ParseException;
 import seedu.edrecord.model.module.Module;
+import seedu.edrecord.model.name.Name;
 import seedu.edrecord.model.person.Address;
 import seedu.edrecord.model.person.Email;
-import seedu.edrecord.model.person.Name;
 import seedu.edrecord.model.person.Phone;
 import seedu.edrecord.model.tag.Tag;
