+++ src/main/java/seedu/edrecord/model/name/Name.java
@@ -1,10 +1,10 @@
-package seedu.edrecord.model.person;
+package seedu.edrecord.model.name;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.edrecord.commons.util.AppUtil.checkArgument;
 
 /**
- * Represents a Person's name in edrecord.
+ * Represents the name of a Person, an Assignment, or a Module in edrecord.
  * Guarantees: immutable; is valid as declared in {@link #isValidName(String)}
  */
 public class Name {
@@ -18,7 +18,7 @@ public class Name {
      */
     public static final String VALIDATION_REGEX = "[\\p{Alnum}][\\p{Alnum} ]*";
 
-    public final String fullName;
+    public final String name;
 
     /**
      * Constructs a {@code Name}.
@@ -28,7 +28,7 @@ public class Name {
     public Name(String name) {
         requireNonNull(name);
         checkArgument(isValidName(name), MESSAGE_CONSTRAINTS);
-        fullName = name;
+        this.name = name;
     }
 
     /**
@@ -41,19 +41,19 @@ public class Name {
 
     @Override
     public String toString() {
-        return fullName;
+        return name;
     }
 
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
                 || (other instanceof Name // instanceof handles nulls
-                && fullName.equals(((Name) other).fullName)); // state check
+                && name.equals(((Name) other).name)); // state check
     }
 
     @Override
     public int hashCode() {
-        return fullName.hashCode();
+        return name.hashCode();
     }
 
 }
