+++ src/test/java/seedu/edrecord/logic/parser/AddCommandParserTest.java
@@ -40,7 +40,7 @@ import seedu.edrecord.logic.commands.AddCommand;
 import seedu.edrecord.model.module.Module;
 import seedu.edrecord.model.person.Address;
 import seedu.edrecord.model.person.Email;
-import seedu.edrecord.model.person.Name;
+import seedu.edrecord.model.name.Name;
 import seedu.edrecord.model.person.Person;
 import seedu.edrecord.model.person.Phone;
 import seedu.edrecord.model.tag.Tag;
