+++ src/test/java/seedu/edrecord/logic/parser/EditCommandParserTest.java
@@ -38,7 +38,7 @@ import seedu.edrecord.logic.commands.EditCommand;
 import seedu.edrecord.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.edrecord.model.person.Address;
 import seedu.edrecord.model.person.Email;
-import seedu.edrecord.model.person.Name;
+import seedu.edrecord.model.name.Name;
 import seedu.edrecord.model.person.Phone;
 import seedu.edrecord.model.tag.Tag;
 import seedu.edrecord.testutil.EditPersonDescriptorBuilder;
